#155: fix circular dependency
Move backfill code into its own file
